Title: a/b 출력, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 250117/a/b 출력/a-divide-b.java
@@ -5,12 +5,12 @@ public class Main {
         // Please write your code here.
         Scanner sc = new Scanner(System.in);
 
-        int a, b, answer;
+        double a, b, answer;
 
         a = sc.nextInt();
         b = sc.nextInt();
 
-        answer = (int)((double)a/b*10000000000)/10000000000.0;
+        answer = (int)(a/b*10000000000)/10000000000.0;
         System.out.print(answer);
     }
 }
